% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gini}
\alias{gini}
\title{Computes the gini coefficient for a given vector.}
\usage{
gini(x, normalize = FALSE)
}
\arguments{
\item{x}{[\code{numeric}]\cr
Numeric vector of quantities.}

\item{normalize}{[\code{logical(1)}]\cr
If \code{TRUE}, the Gini coefficient is normalized. Default is \code{FALSE}.}
}
\value{
[\code{numeric(1)}]
  Gini coefficient.
}
\description{
Computes the gini coefficient for a given vector.
}

