\name{gini}
\alias{gini}
\title{Computes the gini coefficient for a given vector.}
\usage{
gini(x, normalize = FALSE)
}
\arguments{
  \item{x}{[\code{numeric}]\cr Numeric vector of
  quantities.}

  \item{normalize}{[\code{logical(1)}]\cr If \code{TRUE},
  the Gini coefficient is normalized.}
}
\value{
[\code{numeric(1)}] Gini coefficient.
}
\description{
Computes the gini coefficient for a given vector.
}

